+++ Replicator/Pusher.cc
@@ -18,10 +18,8 @@
 //  https://github.com/couchbase/couchbase-lite-core/wiki/Replication-Protocol
 
 #include "Pusher.hh"
-#include "c4BlobStore.h"
 #include "Error.hh"
 #include "StringUtil.hh"
-#include "SecureDigest.hh"
 #include "BLIP.hh"
 #include "HTTPTypes.hh"
 #include <algorithm>
@@ -435,290 +433,7 @@ namespace litecore { namespace repl {
     }
 
 
-#pragma mark - SENDING REVISIONS:
-
-
-    void Pusher::maybeSendMoreRevs() {
-        while (_revisionsInFlight < tuning::kMaxRevsInFlight
-                   && _revisionBytesAwaitingReply <= tuning::kMaxRevBytesAwaitingReply
-                   && !_revsToSend.empty()) {
-            Retained<RevToSend> first = move(_revsToSend.front());
-            _revsToSend.pop_front();
-            sendRevision(first);
-            if (_revsToSend.size() == tuning::kMaxRevsQueued - 1)
-                maybeGetMoreChanges();          // I may now be eligible to send more changes
-        }
-//        if (!_revsToSend.empty())
-//            logVerbose("Throttling sending revs; _revisionsInFlight=%u/%u, _revisionBytesAwaitingReply=%llu/%u",
-//                       _revisionsInFlight, tuning::kMaxRevsInFlight,
-//                       _revisionBytesAwaitingReply, tuning::kMaxRevBytesAwaitingReply);
-    }
-
-    
-    // Send a "rev" message containing a revision body.
-    void Pusher::sendRevision(Retained<RevToSend> rev) {
-        increment(_revisionsInFlight);
-        logVerbose("Sending rev %.*s %.*s (seq #%" PRIu64 ") [%d/%d]",
-                   SPLAT(rev->docID), SPLAT(rev->revID), rev->sequence,
-                   _revisionsInFlight, tuning::kMaxRevsInFlight);
-        sendRevision(rev, [=](MessageProgress progress) {
-            // message progress callback:
-            if (progress.state == MessageProgress::kDisconnected) {
-                doneWithRev(rev, false, false);
-                return;
-            }
-            if (progress.state == MessageProgress::kAwaitingReply) {
-                logDebug("Transmitted 'rev' %.*s #%.*s (seq #%llu)",
-                         SPLAT(rev->docID), SPLAT(rev->revID), rev->sequence);
-                decrement(_revisionsInFlight);
-                increment(_revisionBytesAwaitingReply, progress.bytesSent);
-                maybeSendMoreRevs();
-            }
-            if (progress.state == MessageProgress::kComplete) {
-                decrement(_revisionBytesAwaitingReply, progress.bytesSent);
-                bool synced = !progress.reply->isError();
-                bool completed = true;
-                enum {kNoRetry, kRetryLater, kRetryNow} retry = kNoRetry;
-                if (synced) {
-                    logVerbose("Completed rev %.*s #%.*s (seq #%" PRIu64 ")",
-                               SPLAT(rev->docID), SPLAT(rev->revID), rev->sequence);
-                    finishedDocument(rev);
-                } else {
-                    // Handle an error received from the peer:
-                    auto err = progress.reply->getError();
-                    auto c4err = blipToC4Error(err);
-
-                    if (c4error_mayBeTransient(c4err)) {
-                        completed = false;
-                    } else if (c4err == C4Error{WebSocketDomain, 403}) {
-                        // CBL-123: Retry HTTP forbidden once
-                        if (rev->retryCount++ == 0) {
-                            completed = false;
-                            if (!passive())
-                                retry = kRetryLater;
-                        }
-                    } else if (c4err == C4Error{LiteCoreDomain, kC4ErrorDeltaBaseUnknown}
-                            || c4err == C4Error{LiteCoreDomain, kC4ErrorCorruptDelta}
-                            || c4err == C4Error{WebSocketDomain, int(net::HTTPStatus::UnprocessableEntity)}) {
-                        // CBL-986: On delta error, retry without using delta
-                        if (rev->deltaOK) {
-                            rev->deltaOK = false;
-                            completed = false;
-                            retry = kRetryNow;
-                        }
-                    }
-                    
-                    logError("Got %-serror response to rev '%.*s' #%.*s (seq #%" PRIu64 "): %.*s %d '%.*s'",
-                             (completed ? "" : "transient "),
-                             SPLAT(rev->docID), SPLAT(rev->revID), rev->sequence,
-                             SPLAT(err.domain), err.code, SPLAT(err.message));
-                    finishedDocumentWithError(rev, c4err, !completed);
-                    // If this is a permanent failure, like a validation error or conflict,
-                    // then I've completed my duty to push it.
-                }
-                doneWithRev(rev, completed, synced);
-                switch (retry) {
-                    case kRetryNow:   retryRevs({rev}); break;
-                    case kRetryLater: _revsToRetry.push_back(rev); break;
-                    case kNoRetry:    break;
-                }
-                maybeSendMoreRevs();
-            }
-        });
-    }
-
-
-    void Pusher::couldntSendRevision(RevToSend* rev) {
-        decrement(_revisionsInFlight);
-        doneWithRev(rev, false, false);
-        enqueue(&Pusher::maybeSendMoreRevs);  // async call to avoid recursion
-    }
-
-
-#pragma mark - SENDING ATTACHMENTS:
-
-
-    C4ReadStream* Pusher::readBlobFromRequest(MessageIn *req,
-                                              slice &digestStr,
-                                              Replicator::BlobProgress &progress,
-                                              C4Error *outError)
-    {
-        auto blobStore = _db->blobStore();
-        digestStr = req->property("digest"_sl);
-        progress = {Dir::kPushing};
-        if (!c4blob_keyFromString(digestStr, &progress.key)) {
-            c4error_return(LiteCoreDomain, kC4ErrorInvalidParameter, "Missing or invalid 'digest'"_sl, outError);
-            return nullptr;
-        }
-        int64_t size = c4blob_getSize(blobStore, progress.key);
-        if (size < 0) {
-            c4error_return(LiteCoreDomain, kC4ErrorNotFound, "No such blob"_sl, outError);
-            return nullptr;
-        }
-        progress.bytesTotal = size;
-        return c4blob_openReadStream(blobStore, progress.key, outError);
-    }
-
-
-    // Incoming request to send an attachment/blob
-    void Pusher::handleGetAttachment(Retained<MessageIn> req) {
-        slice digest;
-        Replicator::BlobProgress progress;
-        C4Error err;
-        C4ReadStream* blob = readBlobFromRequest(req, digest, progress, &err);
-        if (blob) {
-            increment(_blobsInFlight);
-            MessageBuilder reply(req);
-            reply.compressed = req->boolProperty("compress"_sl);
-            logVerbose("Sending blob %.*s (length=%" PRId64 ", compress=%d)",
-                       SPLAT(digest), c4stream_getLength(blob, nullptr), reply.compressed);
-            Retained<Replicator> repl = replicator();
-            auto lastNotifyTime = actor::Timer::clock::now();
-            if (progressNotificationLevel() >= 2)
-                repl->onBlobProgress(progress);
-            reply.dataSource = [=](void *buf, size_t capacity) mutable {
-                // Callback to read bytes from the blob into the BLIP message:
-                // For performance reasons this is NOT run on my actor thread, so it can't access
-                // my state directly; instead it calls _attachmentSent() at the end.
-                C4Error err;
-                bool done = false;
-                ssize_t bytesRead = c4stream_read(blob, buf, capacity, &err);
-                progress.bytesCompleted += bytesRead;
-                if (bytesRead < capacity) {
-                    c4stream_close(blob);
-                    this->enqueue(&Pusher::_attachmentSent);
-                    done = true;
-                }
-                if (err.code) {
-                    this->warn("Error reading from blob: %d/%d", err.domain, err.code);
-                    progress.error = {err.domain, err.code};
-                    bytesRead = -1;
-                    done = true;
-                }
-                if (progressNotificationLevel() >= 2) {
-                    auto now = actor::Timer::clock::now();
-                    if (done || now - lastNotifyTime > std::chrono::milliseconds(250)) {
-                        lastNotifyTime = now;
-                        repl->onBlobProgress(progress);
-                    }
-                }
-                return (int)bytesRead;
-            };
-            req->respond(reply);
-            return;
-        }
-        req->respondWithError(c4ToBLIPError(err));
-    }
-
-
-    void Pusher::_attachmentSent() {
-        decrement(_blobsInFlight);
-    }
-
-
-    // Incoming request to prove I have an attachment that I'm pushing, without sending it:
-    void Pusher::handleProveAttachment(Retained<MessageIn> request) {
-        slice digest;
-        Replicator::BlobProgress progress;
-        C4Error err;
-        c4::ref<C4ReadStream> blob = readBlobFromRequest(request, digest, progress, &err);
-        if (blob) {
-            logVerbose("Sending proof of attachment %.*s", SPLAT(digest));
-            SHA1Builder sha;
-
-            // First digest the length-prefixed nonce:
-            slice nonce = request->body();
-            if (nonce.size == 0 || nonce.size > 255) {
-                request->respondWithError({"BLIP"_sl, 400, "Missing nonce"_sl});
-                return;
-            }
-            sha << (nonce.size & 0xFF) << nonce;
-
-            // Now digest the attachment itself:
-            static constexpr size_t kBufSize = 8192;
-            auto buf = make_unique<uint8_t[]>(kBufSize);
-            size_t bytesRead;
-            while ((bytesRead = c4stream_read(blob, buf.get(), kBufSize, &err)) > 0)
-                sha << slice(buf.get(), bytesRead);
-            buf.reset();
-            blob = nullptr;
-            
-            if (err.code == 0) {
-                // Respond with the base64-encoded digest:
-                C4BlobKey proofDigest;
-                sha.finish(proofDigest.bytes, sizeof(proofDigest.bytes));
-                alloc_slice proofStr = c4blob_keyToString(proofDigest);
-
-                MessageBuilder reply(request);
-                reply.write(proofStr);
-                request->respond(reply);
-                return;
-            }
-        }
-
-        // If we got here, we failed:
-        request->respondWithError(c4ToBLIPError(err));
-    }
-
-
 #pragma mark - PROGRESS:
-
-
-    void Pusher::doneWithRev(RevToSend *rev, bool completed, bool synced) {
-        if (!passive()) {
-            addProgress({rev->bodySize, 0});
-            if (completed) {
-                _checkpointer.completedSequence(rev->sequence);
-
-                auto lastSeq = _checkpointer.localMinSequence();
-                if (lastSeq / 1000 > _lastSequenceLogged / 1000 || willLog(LogLevel::Verbose))
-                    logInfo("Checkpoint now %s", _checkpointer.to_string().c_str());
-                _lastSequenceLogged = lastSeq;
-            }
-        }
-
-        if (synced && _options.push > kC4Passive)
-            _db->markRevSynced(const_cast<RevToSend*>(rev));
-
-        // Remove rev from _pushingDocs, and see if there's a newer revision to send next:
-        auto i = _pushingDocs.find(rev->docID);
-        if (i == _pushingDocs.end()) {
-            if (connected())
-                warn("_donePushingRev('%.*s'): That docID is not active!", SPLAT(rev->docID));
-            return;
-        }
-
-        Retained<RevToSend> newRev = i->second;
-        _pushingDocs.erase(i);
-        if (newRev) {
-            if (synced && getForeignAncestors())
-                newRev->remoteAncestorRevID = rev->revID;
-            logVerbose("Now that '%.*s' %.*s is done, propose %.*s (remote %.*s) ...",
-                       SPLAT(rev->docID), SPLAT(rev->revID), SPLAT(newRev->revID),
-                       SPLAT(newRev->remoteAncestorRevID));
-            bool ok = false;
-            if (synced && getForeignAncestors()
-                && c4rev_getGeneration(newRev->revID) <= c4rev_getGeneration(rev->revID)) {
-                // Don't send; it'll conflict with what's on the server
-            } else {
-                // Send newRev as though it had just arrived:
-                bool should = _db->use<bool>([&](C4Database *db) {
-                    return shouldPushRev(newRev, nullptr, db);
-                });
-                if (should) {
-                    _maxPushedSequence = max(_maxPushedSequence, rev->sequence);
-                    gotOutOfOrderChange(newRev);
-                    ok = true;
-                }
-            }
-            if (!ok) {
-                logVerbose("   ... nope, decided not to propose '%.*s' %.*s",
-                           SPLAT(newRev->docID), SPLAT(newRev->revID));
-            }
-        } else {
-            logDebug("Done pushing '%.*s' %.*s", SPLAT(rev->docID), SPLAT(rev->revID));
-        }
-    }
 
 
     void Pusher::_connectionClosed() {
